src/test/java/com/example/demo/service/CityServiceTest.java
package com.example.demo.service;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.junit.Assert;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;
import static org.junit.Assert.*;

@RunWith(SpringRunner.class)
@SpringBootTest
@AutoConfigureMockMvc
public class CityServiceTest {

    @Autowired
    private CityService cityService;

    @Test
    public void delete() throws Exception{
        int i = cityService.delete(2);
        Assert.assertEquals(1, i);
    }
}